EngineController Major Updates
Major updates to my Engine Controller Software. Contains controls for
solenoid as well as servo valves for my liquid fueled igniter and
engine. Measurement software for psi transducers, load cells, and
thermocouples has been updated or added. Improved safety features such
as overrides are now available. For complete documentation review the
source code.

Note the main library is EngineController.ino and can be found in the
directory EngineController off of root. Load this onto an arduino
microcontroller as a project. Dependent libraries probably need to be
moved to the libraries directory of the arduino installation.